# backend/docs/swagger.yaml
definitions:
...
  github_com_raids-lab_orbit_internal_resputil.ErrorCode:
    enum:
    - 0
    - 40001
    - 40002
    - 40101
    - 40102
    - 40111
    - 40112
    - 40113
    - 40114
    - 40121
    - 40122
    - 40123
    - 40301
    - 40302
    - 40401
    - 50001
    - 99999
    type: integer
    x-enum-varnames:
    - OK
    - InvalidRequest
    - BusinessLogicError
    - TokenExpired
    - TokenInvalid
    - InvalidCredentials
    - LdapError
    - LdapUserNotFound
    - LegacyTokenNotSupported
    - MustRegister
    - UidServiceError
    - UidNotFound
    - UserNotAllowed
    - UserEmailNotVerified
    - ServiceSshdNotFound
    - ServiceError
    - NotSpecified
  github_com_raids-lab_orbit_internal_resputil.List-internal_handler_operations_OperationLogResp:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_handler_operations.OperationLogResp'
        type: array
      total:
        type: integer
    type: object
  github_com_raids-lab_orbit_internal_resputil.Response-any:
    properties:
      code:
        $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.ErrorCode'
      data: {}
      msg:
        type: string
    type: object
...
  ? github_com_raids-lab_orbit_internal_resputil.Response-github_com_raids-lab_orbit_internal_resputil_List-internal_handler_operations_OperationLogResp
  : properties:
      code:
        $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.ErrorCode'
      data:
        $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.List-internal_handler_operations_OperationLogResp'
      msg:
        type: string
    type: object
...
  github_com_raids-lab_orbit_internal_resputil.Response-map_string_string:
    properties:
      code:
        $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.ErrorCode'
      data:
        $ref: '#/definitions/map_string_string'
      msg:
        type: string
    type: object
  github_com_raids-lab_orbit_internal_resputil.Response-string:
    properties:
      code:
        $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.ErrorCode'
      data:
        type: string
      msg:
        type: string
    type: object
...
  internal_handler_operations.OperationLogResp:
    properties:
      created_at:
        type: string
      details:
        type: object
      error_message:
        type: string
      id:
        type: integer
      operation_type:
        type: string
      operator:
        type: string
      operator_role:
        type: string
      status:
        type: string
      target:
        type: string
      updated_at:
        type: string
    type: object
...
  map_string_string:
    additionalProperties:
      type: string
    type: object
...
info:
  contact: {}
  description: This is the API server for Orbit, a Multi-tenant AI Model Training
    Platform based on Kubernetes.
  title: Orbit API
  version: 1.0.0
paths:
...
  /v1/admin/operation-logs:
    delete:
      consumes:
      - application/json
      description: 删除所有操作日志记录，仅用于测试/调试场景
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-map_string_string'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 清空操作日志
      tags:
      - OperationLog
    get:
      consumes:
      - application/json
      description: 获取操作日志列表，支持分页和筛选
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 操作类型
        in: query
        name: operation_type
        type: string
      - description: 操作人
        in: query
        name: operator
        type: string
      - description: 操作对象
        in: query
        name: target
        type: string
      - description: 模糊搜索关键词
        in: query
        name: search
        type: string
      - description: 开始时间，格式例如：2024-01-02T15:04:05Z
        in: query
        name: start_time
        type: string
      - description: 结束时间，格式例如：2024-01-02T15:04:05Z
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-github_com_raids-lab_orbit_internal_resputil_List-internal_handler_operations_OperationLogResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 获取操作日志列表
      tags:
      - OperationLog
...
  /v1/models/downloads/{id}/logs:
    get:
      consumes:
      - application/json
      description: 获取模型下载任务的实时日志
      parameters:
      - description: 下载任务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-string'
      security:
      - Bearer: []
      summary: 获取模型下载任务的 Pod 日志
      tags:
      - ModelDownload
...
  /v1/namespaces/{namespace}/pods/{name}/containers/{container}/log:
    get:
      consumes:
      - application/json
      description: 获取Pod容器日志
      parameters:
      - description: 命名空间
        in: path
        name: namespace
        required: true
        type: string
      - description: Pod名称
        in: path
        name: name
        required: true
        type: string
      - description: 容器名称
        in: path
        name: container
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Pod容器日志
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "400":
          description: Request parameter error
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
        "500":
          description: Other errors
          schema:
            $ref: '#/definitions/github_com_raids-lab_orbit_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 获取Pod容器日志
      tags:
      - Pod
...
securityDefinitions:
  Bearer:
    description: 访问 /login 并获取 TOKEN 后，填入 'Bearer ${TOKEN}' 以访问受保护的接口
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"